fix: rename tools/ to _benchmarks/ to avoid namespace collision with hermes-agent
The tools/ directory contains benchmark scripts that have no __init__.py,
making them a namespace package. When installed into site-packages, this
shadows hermes-agent's own tools/ package, causing ImportError on
from tools import check_file_requirements and silently stripping
file tools (read_file, write_file, patch, search_files).

Renaming to _benchmarks/ eliminates the collision. Updated all imports.
