+++ _benchmarks/bench_nai0.py
