+++ _benchmarks/bench_nai1.py
