+++ _benchmarks/diagnostic_tr.py
