+++ _benchmarks/generate_beam_charts.py
